Move main to separate bin folder, add generate subcommand
