+++ modmod/src/patch.rs
@@ -47,7 +47,7 @@ impl Patch {
                 .unwrap();
 
             let mut new_file = new_file_path.open_file()?;
-            let old_file_path = old_dir.as_ref().join(&relative_file_path);
+            let old_file_path = old_dir.as_ref().join(relative_file_path);
 
             let mut old_file = match fs::File::open(&old_file_path) {
                 Ok(f) => f,
